Updated the readme file
I changed the title for more clarity on the project. This will help users have an idea of the repo before clicking on it. I also added some headers to help organize the ideas written. I modified some of the text to make it more generalized so more users can contribute to the project. I also added a list of commands that were already added to allow users to access the exact interface or system they need. This list will also help users who want to contribute know if their interface or system is already added and where to branch to.